v1.0.5
- Added login, password and host get / set commands
- Corrected a bug in the API where a Long was used instead of a Double
- Added reboot command
- Added read marker update api call

+++ src/main/java/io/github/ma1uta/matrix/client/methods/async/TagAsyncMethods.java
@@ -62,7 +62,7 @@ public class TagAsyncMethods {
      * @param order  The tag order.
      * @return The empty response.
      */
-    public CompletableFuture<EmptyResponse> add(String roomId, String tag, Long order) {
+    public CompletableFuture<EmptyResponse> add(String roomId, String tag, Double order) {
         String userId = connectionInfo.getUserId();
         Objects.requireNonNull(roomId, "RoomId cannot be empty.");
         Objects.requireNonNull(userId, "UserId cannot be empty.");
