+++ src/main/java/io/github/ma1uta/matrix/client/methods/blocked/TagMethods.java
@@ -61,7 +61,7 @@ public class TagMethods {
      * @param order  The tag order.
      * @return The empty response.
      */
-    public EmptyResponse add(String roomId, String tag, Long order) {
+    public EmptyResponse add(String roomId, String tag, Double order) {
         String userId = connectionInfo.getUserId();
         Objects.requireNonNull(roomId, "RoomId cannot be empty.");
         Objects.requireNonNull(userId, "UserId cannot be empty.");
